Заменен старый код отправки email уведомлений напрямую на использование асинхронной задачи Celery. правки
