+++ documents/tasks.py
@@ -1,23 +1,13 @@
 from celery import shared_task
 from django.core.mail import send_mail
 from django.conf import settings
-from django.contrib.auth.models import Group
+from documents.services import get_admin_emails
+from users.models import User
 from documents.models import Document
 
 
 @shared_task
-def get_admin_emails(group_name='Document Administrators'):
-    """
-    Возвращает список email адресов администраторов из указанной группы.
-    """
-    try:
-        group = Group.objects.get(name=group_name)
-        return [user.email for user in group.user_set.filter(mailing=True)]
-    except Group.DoesNotExist:
-        return []
-
-
-def send_document_creation_email_task(document_id, request_url):
+def send_document_creation_email(document_id, request_url):
     """
     Асинхронная задача для отправки email уведомления о создании нового документа.
     """
@@ -32,8 +22,7 @@ def send_document_creation_email_task(document_id, request_url):
         f"Просмотреть документ: {request_url}"
     )
 
-    admin_emails = [user.email for user in
-                    Group.objects.get(name='Document Administrators').user_set.filter(mailing=True)]
+    admin_emails = get_admin_emails()
 
     send_mail(
         subject,
@@ -49,9 +38,6 @@ def send_document_status_email(user_id, document_id, status, comment=None):
     """
         Отправляет email уведомление пользователю о статусе его документа.
     """
-    from users.models import User
-    from documents.models import Document
-
     user = User.objects.get(id=user_id)
     document = Document.objects.get(id=document_id)
